TWINE: moved file into subdir
