+++ engines/twine/renderer/renderer.cpp
@@ -25,12 +25,12 @@
 #include "common/stream.h"
 #include "common/textconsole.h"
 #include "common/util.h"
-#include "twine/scene/actor.h"
 #include "twine/menu/interface.h"
 #include "twine/menu/menu.h"
-#include "twine/scene/movements.h"
 #include "twine/renderer/redraw.h"
-#include "twine/shadeangletab.h"
+#include "twine/renderer/shadeangletab.h"
+#include "twine/scene/actor.h"
+#include "twine/scene/movements.h"
 #include "twine/shared.h"
 #include "twine/twine.h"
