+++ engines/twine/renderer/shadeangletab.h
